+++ docs/make.jl
@@ -36,6 +36,7 @@ makedocs(
             "cli/generate.md",
             "cli/marginalize.md",
             "cli/polish.md",
+            "cli/version.md",
         ],
         "Development" => [
             "dev/building-docker.md",